Create sub mod
