+++ nym-vpn-core/crates/nym-vpn-lib/src/platform/mod.rs
@@ -8,11 +8,13 @@ pub(crate) mod error;
 #[cfg(any(target_os = "ios", target_os = "macos"))]
 pub mod swift;
 
-use std::{env, path::PathBuf, str::FromStr, sync::Arc};
+mod account;
 
+use std::{env, path::PathBuf, sync::Arc};
+
+use account::AccountControllerHandle;
 use lazy_static::lazy_static;
 use log::*;
-use nym_vpn_account_controller::{AccountCommand, ReadyToConnect, SharedAccountState};
 use tokio::{
     runtime::Runtime,
     sync::{mpsc, Mutex},
@@ -22,7 +24,6 @@ use tokio_util::sync::CancellationToken;
 use url::Url;
 
 use nym_gateway_directory::Config as GatewayDirectoryConfig;
-use nym_vpn_store::mnemonic::MnemonicStorage as _;
 
 use self::error::VpnError;
 #[cfg(target_os = "android")]
@@ -59,7 +60,7 @@ async fn start_vpn_inner(config: VPNConfig) -> Result<(), VpnError> {
     // We want to move this check into the state machine so that it happens during the connecting
     // state instead. This would allow us more flexibility in waiting for the account to be ready
     // and handle errors in a unified manner.
-    assert_account_ready_to_connect().await?;
+    account::assert_account_ready_to_connect().await?;
 
     let mut guard = STATE_MACHINE_HANDLE.lock().await;
 
@@ -99,41 +100,15 @@ async fn stop_vpn_inner() -> Result<(), VpnError> {
 #[allow(non_snake_case)]
 #[uniffi::export]
 pub fn startAccountController(data_dir: String) -> Result<(), VpnError> {
-    RUNTIME.block_on(start_account_controller_inner(PathBuf::from(data_dir)))
-}
-
-async fn start_account_controller_inner(data_dir: PathBuf) -> Result<(), VpnError> {
-    let mut guard = ACCOUNT_CONTROLLER_HANDLE.lock().await;
-
-    if guard.is_none() {
-        let account_controller_handle = start_account_controller(data_dir).await?;
-        *guard = Some(account_controller_handle);
-        Ok(())
-    } else {
-        Err(VpnError::InvalidStateError {
-            details: "Account controller is already running.".to_owned(),
-        })
-    }
+    RUNTIME.block_on(account::start_account_controller_inner(PathBuf::from(
+        data_dir,
+    )))
 }
 
 #[allow(non_snake_case)]
 #[uniffi::export]
 pub fn stopAccountController() -> Result<(), VpnError> {
-    RUNTIME.block_on(stop_account_controller_inner())
-}
-
-async fn stop_account_controller_inner() -> Result<(), VpnError> {
-    let mut guard = ACCOUNT_CONTROLLER_HANDLE.lock().await;
-
-    match guard.take() {
-        Some(account_controller_handle) => {
-            account_controller_handle.shutdown_and_wait().await;
-            Ok(())
-        }
-        None => Err(VpnError::InvalidStateError {
-            details: "Account controller is not running.".to_owned(),
-        }),
-    }
+    RUNTIME.block_on(account::stop_account_controller_inner())
 }
 
 #[allow(non_snake_case)]
@@ -163,97 +138,28 @@ async fn fetch_environment(network_name: &str) -> Result<NetworkEnvironment, Vpn
         })
 }
 
-fn setup_account_storage(path: &str) -> Result<crate::storage::VpnClientOnDiskStorage, VpnError> {
-    let path = PathBuf::from_str(path).map_err(|err| VpnError::InternalError {
-        details: err.to_string(),
-    })?;
-    Ok(crate::storage::VpnClientOnDiskStorage::new(path))
-}
-
 #[allow(non_snake_case)]
 #[uniffi::export]
 pub fn storeAccountMnemonic(mnemonic: String, path: String) -> Result<(), VpnError> {
-    RUNTIME.block_on(store_account_mnemonic(&mnemonic, &path))
-}
-
-async fn store_account_mnemonic(mnemonic: &str, path: &str) -> Result<(), VpnError> {
-    // TODO: store the mnemonic by sending a command to the account controller instead of directly
-    // interacting with the storage.
-
-    let storage = setup_account_storage(path)?;
-
-    let mnemonic = nym_vpn_store::mnemonic::Mnemonic::parse(mnemonic).map_err(|err| {
-        VpnError::InternalError {
-            details: err.to_string(),
-        }
-    })?;
-
-    storage
-        .store_mnemonic(mnemonic)
-        .await
-        .map_err(|err| VpnError::InternalError {
-            details: err.to_string(),
-        })?;
-
-    send_account_command(AccountCommand::UpdateSharedAccountState).await?;
-
-    Ok(())
+    RUNTIME.block_on(account::store_account_mnemonic(&mnemonic, &path))
 }
 
 #[allow(non_snake_case)]
 #[uniffi::export]
 pub fn isAccountMnemonicStored(path: String) -> Result<bool, VpnError> {
-    RUNTIME.block_on(is_account_mnemonic_stored(&path))
-}
-
-async fn is_account_mnemonic_stored(path: &str) -> Result<bool, VpnError> {
-    // TODO: query the mnemonic by sending a command to the account controller instead of directly
-    // interacting with the storage.
-
-    let storage = setup_account_storage(path)?;
-    storage
-        .is_mnemonic_stored()
-        .await
-        .map_err(|err| VpnError::InternalError {
-            details: err.to_string(),
-        })
+    RUNTIME.block_on(account::is_account_mnemonic_stored(&path))
 }
 
 #[allow(non_snake_case)]
 #[uniffi::export]
 pub fn removeAccountMnemonic(path: String) -> Result<bool, VpnError> {
-    RUNTIME.block_on(remove_account_mnemonic(&path))
-}
-
-async fn remove_account_mnemonic(path: &str) -> Result<bool, VpnError> {
-    // TODO: remove the mnemonic by sending a command to the account controller instead of directly
-    // interacting with the storage.
-
-    let storage = setup_account_storage(path)?;
-    let is_account_removed_success =
-        storage
-            .remove_mnemonic()
-            .await
-            .map(|_| true)
-            .map_err(|err| VpnError::InternalError {
-                details: err.to_string(),
-            })?;
-
-    send_account_command(AccountCommand::UpdateSharedAccountState).await?;
-
-    Ok(is_account_removed_success)
+    RUNTIME.block_on(account::remove_account_mnemonic(&path))
 }
 
 #[allow(non_snake_case)]
 #[uniffi::export]
 pub fn getAccountSummary() -> Result<AccountStateSummary, VpnError> {
-    RUNTIME.block_on(get_account_summary())
-}
-
-async fn get_account_summary() -> Result<AccountStateSummary, VpnError> {
-    let shared_account_state = get_shared_account_state().await?;
-    let account_state_summary = shared_account_state.lock().await.clone();
-    Ok(AccountStateSummary::from(account_state_summary))
+    RUNTIME.block_on(account::get_account_summary())
 }
 
 #[allow(non_snake_case)]
@@ -351,94 +257,6 @@ pub struct VPNConfig {
 #[uniffi::export(with_foreign)]
 pub trait TunnelStatusListener: Send + Sync {
     fn on_event(&self, event: TunnelEvent);
-}
-
-struct AccountControllerHandle {
-    command_sender: mpsc::UnboundedSender<AccountCommand>,
-    shared_state: nym_vpn_account_controller::SharedAccountState,
-    handle: JoinHandle<()>,
-    shutdown_token: CancellationToken,
-}
-
-impl AccountControllerHandle {
-    fn send_command(&self, command: AccountCommand) {
-        if let Err(e) = self.command_sender.send(command) {
-            tracing::error!("Failed to send comamnd: {}", e);
-        }
-    }
-
-    async fn shutdown_and_wait(self) {
-        self.shutdown_token.cancel();
-
-        if let Err(e) = self.handle.await {
-            tracing::error!("Failed to join on account controller handle: {}", e);
-        }
-    }
-}
-
-async fn start_account_controller(data_dir: PathBuf) -> Result<AccountControllerHandle, VpnError> {
-    let storage = Arc::new(tokio::sync::Mutex::new(
-        crate::storage::VpnClientOnDiskStorage::new(data_dir.clone()),
-    ));
-    // TODO: pass in as argument
-    let user_agent = crate::util::construct_user_agent();
-    let shutdown_token = CancellationToken::new();
-    let account_controller = nym_vpn_account_controller::AccountController::new(
-        Arc::clone(&storage),
-        data_dir.clone(),
-        user_agent,
-        shutdown_token.child_token(),
-    )
-    .await;
-
-    let shared_account_state = account_controller.shared_state();
-    let account_command_tx = account_controller.command_tx();
-    let account_controller_handle = tokio::spawn(account_controller.run());
-
-    Ok(AccountControllerHandle {
-        command_sender: account_command_tx,
-        shared_state: shared_account_state,
-        handle: account_controller_handle,
-        shutdown_token,
-    })
-}
-
-async fn send_account_command(command: AccountCommand) -> Result<(), VpnError> {
-    if let Some(guard) = &*ACCOUNT_CONTROLLER_HANDLE.lock().await {
-        guard.send_command(command);
-        Ok(())
-    } else {
-        Err(VpnError::InvalidStateError {
-            details: "Account controller is not running.".to_owned(),
-        })
-    }
-}
-
-async fn get_shared_account_state() -> Result<SharedAccountState, VpnError> {
-    if let Some(guard) = &*ACCOUNT_CONTROLLER_HANDLE.lock().await {
-        Ok(guard.shared_state.clone())
-    } else {
-        Err(VpnError::InvalidStateError {
-            details: "Account controller is not running.".to_owned(),
-        })
-    }
-}
-
-async fn is_account_ready_to_connect() -> Result<ReadyToConnect, VpnError> {
-    Ok(get_shared_account_state()
-        .await?
-        .is_ready_to_connect()
-        .await)
-}
-
-async fn assert_account_ready_to_connect() -> Result<(), VpnError> {
-    match is_account_ready_to_connect().await? {
-        ReadyToConnect::Ready => Ok(()),
-        not_ready_to_connect => {
-            tracing::warn!("Not ready to connect: {:?}", not_ready_to_connect);
-            Err(VpnError::Account(not_ready_to_connect.into()))
-        }
-    }
 }
 
 struct StateMachineHandle {
